Revert "Gcp sub feature add security"
