+++ src/main/java/com/digiassist/ticket/TicketApplication.java
@@ -3,7 +3,7 @@ package com.digiassist.ticket;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-@SpringBootApplication(scanBasePackages = {"com.digiassist.auth", "com.digiassist.ticket"})
+@SpringBootApplication
 public class TicketApplication {
 
 	public static void main(String[] args) {
