+++ src/main/resources/application.properties
@@ -1,8 +1,4 @@
-#spring.webflux.base-path=/tickets
-server.servlet.context-path=/tickets
-#spring.cloud.gcp.firestore.project-id=gcms-oht29077u9-2022
-spring.main.allow-circular-references=true
-spring.mvc.pathmatch.matching-strategy = ANT_PATH_MATCHER
+spring.webflux.base-path=/tickets
 
 #spring.cloud.gcp.firestore.credentials.location=digifire-93126cbbfbc0.json
 #spring.cloud.gcp.firestore.project-id=digifire
