+++ src/test/java/com/digiassist/ticket/TicketApplicationTests.java
@@ -1,4 +1,3 @@
-
 package com.digiassist.ticket;
 
 import org.junit.jupiter.api.Test;
